Most of Milestone 3 complete
Added Enemies that move around and follow the player

The enemies use a nav mesh so they can path find to the player's location

Rearranged the inheritance tree. The player and enemies are characters, and all characters are weapon users.

Added a wave system for the bonus feature

Enemies use weapons to attack the player